database.models -> 도메인별로 이동; user 앱 리크루팅에 맞게 수정
