+++ wacruit/src/apps/problem/models.py
@@ -8,11 +8,11 @@ from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import mapped_column
 from sqlalchemy.orm import relationship
 
-from wacruit.src.database.models.base import DeclarativeBase
-from wacruit.src.database.models.base import intpk
+from wacruit.src.database.base import DeclarativeBase
+from wacruit.src.database.base import intpk
 
 if TYPE_CHECKING:
-    from wacruit.src.database.models import User
+    from wacruit.src.apps.user.models import User
 
 
 class Problem(DeclarativeBase):
