+++ wacruit/src/apps/user/repositories.py
@@ -1,9 +1,9 @@
 from fastapi import Depends
 from sqlalchemy.orm import Session
 
+from wacruit.src.apps.user.models import User
 from wacruit.src.database.connection import get_db_session
 from wacruit.src.database.connection import Transaction
-from wacruit.src.database.models import User  # type: ignore
 
 
 class UserRepository:
