+++ wacruit/src/apps/user/views.py
@@ -7,26 +7,29 @@ from fastapi import HTTPException
 from fastapi import Security
 from fastapi.security import APIKeyHeader
 
+from wacruit.src.apps.user.models import User
 from wacruit.src.apps.user.repositories import UserRepository
 from wacruit.src.apps.user.schemas import UserCreateUpdateRequest
 from wacruit.src.apps.user.services import UserService
-from wacruit.src.database.models.user import User
 
 v1_router = APIRouter(prefix="/v1/users", tags=["users"])
 
 
 def get_current_user(
-    waffle_user_id: str = Security(
-        APIKeyHeader(
-            name="waffle-user-id",
-            scheme_name="waffle-user-id",
-            description=(
-                "와플스튜디오 SSO를 통해 발급받은 액세스 토큰에 포함된 사용자의 고유 식별자입니다. "
-                "액세스 토큰을 디코드하면 확인할 수 있습니다."
-            ),
-        )
-    ),
-    user_repository: UserRepository = Depends(),
+    waffle_user_id: Annotated[
+        str,
+        Security(
+            APIKeyHeader(
+                name="waffle-user-id",
+                scheme_name="waffle-user-id",
+                description=(
+                    "와플스튜디오 SSO를 통해 발급받은 액세스 토큰에 포함된 사용자의 고유 식별자입니다. "
+                    "액세스 토큰을 디코드하면 확인할 수 있습니다."
+                ),
+            )
+        ),
+    ],
+    user_repository: Annotated[UserRepository, Depends()],
 ) -> User:
     user = user_repository.get_user_by_sso_id(waffle_user_id)
     if user is None:
@@ -38,7 +41,7 @@ def get_current_user(
 def create_user(
     request: UserCreateUpdateRequest,
     waffle_user_id: Annotated[str, Header()],
-    user_service: UserService = Depends(),
+    user_service: Annotated[UserService, Depends()],
 ):
     request.sso_id = waffle_user_id
     return user_service.create_user(request)
@@ -46,7 +49,7 @@ def create_user(
 
 @v1_router.get("")
 def list_users(
-    user_service: UserService = Depends(),
+    user_service: Annotated[UserService, Depends()],
     # current_user: User = Depends(get_current_user),
 ):
     # if not current_user.is_admin:
