+++ wacruit/src/database/base.py
@@ -9,7 +9,4 @@ DeclarativeBase: Type[Base] = declarative_base()
 
 intpk = Annotated[int, mapped_column(primary_key=True, autoincrement=True)]
 str30 = Annotated[str, mapped_column(String(30))]
-str50_default_none = Annotated[
-    str | None,
-    mapped_column(String(50), nullable=True, default=None),
-]
+str50 = Annotated[str, mapped_column(String(50))]
