--- wacruit/src/database/models/README.md
@@ -1,3 +1,0 @@
-# Deprecated Warning
-
-이 폴더의 파일은 deprecated 되었습니다. `wacruit/src/apps` 의 `models` 폴더 또는 파일에 ORM 모델을 정의해주세요.
